Better way to WAR the Eigen issue
Instead of having the CI script bump Eigen up to master, just change
the referenced submodule commit in our 16.11 repository.

**WARNING** This change apparently can't be upstreamed as-is because
of problems on Android with Eigen master.  But it's fine for us, as
our 16.11 container won't run on Android.  And it will be easier to
recall the discrepancy between upstream and our branch if this isn't
buried in the CI script.  Plus this way 'docker build .' from the
local system as opposed to via the CI runner does more like what you'd
expect.