Improve dagnet chain computation by pruning redundant dependencies
Summary:
We have noticed that the number of chains computed is usually much larger than necessary, when there is a backward pass. For example having a network of 5 FCs with gradient operators (but no parameter updates) should yield only one chain, but instead over 20 were created.  After adding parameter updates, the forward pass still should remain one chain, while the backward pass will be splintered.

Analysis showed that the problem was the dependices from forward ops to the gradient computation. But these are redundant since the gradient op is already dependent from the op via the full path over ops. Example:

  fc1     -> fc2   --->   fc3  --> loss
    |          |             |          |
  fc1grad <- fc2grad    <- fc3grad <-

Here fc1 and fc1 grad have a direct dependency, but indirect dependency via fc2->fc3->[...]->fc1grad already covers that dependency.

To fix this, I added a pruning step prior to the chain computation. The chain computation is done on the pruned tree, but I do not modify the runtime chains for safety.

Pruning is based on  following logic:
  - if one of my direct parents is ancestor via an another traversar, I can remove the direct dependency

Pruning is extremely fast, linear in the number of dependencies.

Reviewed By: dzhulgakov

Differential Revision: D4500293

fbshipit-source-id: 0994ae6775c53378ea1e0074365cef041764a1b4