Default LocalSession to current workspace.
Summary:
At the moment LocalSession creates a new workspace if none if provided. As a
result anything that have been executed in local session is not going to be
avaiable to the external caller, i.e. everything that is using SingleRunner can
only observe side-effects and not really access intermediate blobs.

This diff is modifying LocalSession to run in current workspace instead (unless
it gots some really weird effects because we rely on privateness of the
workspace it should work).

Differential Revision: D4634743

fbshipit-source-id: 975bed154c7ca215dc3fc0d60f05a7c092711482