Update CentOS installation guide. (#502)
* Update CentOS installation guide.

The gflags provided in yum of the current CentOS does not match the dependency requirement of the latest glog version. It also need to be cloned from Github and manually installed from source. Also, the glog need the latest CMake3 to reconfigure.

* Fix gflags and glog installing scripts.

Add "-fPIC" to compilation flags.
Do not proceed on any failure.