Require C++11 support with CMake functions.
This PR replaces PR #464. It requires C+11 support using the
new CMake variables (CMAKE_CXX_STANDARD, CMAKE_CXX_STANDARD_REQUIRED,
etc.) when CMake is version 3.1 or above. Otherwise, if CMake is older
(e.g. Ubuntu 14.04) it falls back to using the -std=c++11 flag and
issues a warning.

This PR is based on the comment from @Yangqing:
https://github.com/caffe2/caffe2/pull/464#issuecomment-305376923

The corresponding line in cmake/MiscCheck.cmake is removed in order to
reduce redundancy.