Improve logic when creating a common world
Summary:
When creating a common world, we would attempt to create
one using an existing common world to save on setup cost. This could cause
unexpected behavior when the backing common world had a shorter
timeout than the world being created. This patch improves this
logic by limiting the usage of a backing world to only ones that
have a long enough timeout.

Reviewed By: andrewwdye

Differential Revision: D5570904

fbshipit-source-id: d3b5073a64381ed068a30dcc461a6ec9ce15ad9c