handle reshape gradient in shape inference in special way
Summary:
Reshape op's gradient op will have the original shape stored in a blob. Shape inference won't work directly because shape inference function does not have access to blob contents.
In this case, I think making a special exception in the shape inference system is justified: we store the output of reshape in a reshape-cache, and pass that in a backward pass.

Also include my experimental test script that I used for NeuralMT CNN model.

Reviewed By: asaadaldien

Differential Revision: D5721502

fbshipit-source-id: fdc8ab901d3bee2c4621ee5140a5435e49f4471d