Fixes for NCCLReduce with non-zero root
All other NCCL ops expect paired src, dst pointers for each
GPU. Reduce doesn't, and the old logic would always set dst for
rank = 0 regardless of whether that was the root or not.
This change takes into account that Reduce only has one output, and it
should assign dst only for the root rank. Also changes the schema to
allow inplace for any input and Output(0).